refactor: decompose welcome message
