+++ src/components/chat/message/Message.scss
@@ -1,4 +1,4 @@
-@import '../../scss/variables.scss';
+@import '../../../scss/variables.scss';
 
 .common_date {
   text-align: center;
